+++ src/api/agent-auth.ts
@@ -1,7 +1,7 @@
 import * as AsyncLock from "async-lock";
 import * as debug from "debug";
 import * as jwt from "jsonwebtoken";
-import fetch from "node-fetch";
+import fetch from "cross-fetch";
 import "url-search-params-polyfill";
 import * as uuid from "uuid";
 import {
@@ -94,7 +94,7 @@ export abstract class AgentAuth extends MindConnectBase implements TokenRotation
                 body: JSON.stringify(body),
                 headers: headers,
                 agent: this._proxyHttpAgent,
-            });
+            } as RequestInit);
 
             if (!response.ok) {
                 throw new Error(`${response.statusText} ${await response.text()}`);
@@ -197,7 +197,7 @@ export abstract class AgentAuth extends MindConnectBase implements TokenRotation
                 body: JSON.stringify(body),
                 headers: headers,
                 agent: this._proxyHttpAgent,
-            });
+            } as RequestInit);
 
             if (!response.ok) {
                 throw new Error(`${response.statusText} ${await response.text()}`);
@@ -305,7 +305,7 @@ export abstract class AgentAuth extends MindConnectBase implements TokenRotation
                 body: body,
                 headers: headers,
                 agent: this._proxyHttpAgent,
-            });
+            } as RequestInit);
 
             if (!response.ok) {
                 throw new Error(`${response.statusText} ${await response.text()}`);
@@ -356,7 +356,7 @@ export abstract class AgentAuth extends MindConnectBase implements TokenRotation
                 method: "GET",
                 headers: headers,
                 agent: this._proxyHttpAgent,
-            });
+            } as RequestInit);
 
             if (!response.ok) {
                 throw new Error(`${response.statusText} ${await response.text()}`);