+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - CLI: mc dev-proxy added simple api proxy for mindsphere local app development
 - CLI: deprecated use of classical service credentials
 - Improved documentation for GetDataSourceConfiguration and GetDataMappings methods (#149, #150)
+- Switched the fetch library to cross-fetch for browser support for SDK
 - bumped all dependencies
 
 ## 3.7.0 - (Boysenberry Vienna) - February 2020